feature:simplified layout file, update test files
